+++ client/src/public/ts/socket.ts
@@ -5,7 +5,7 @@ import {
 } from "./index";
 import { updateStatusTrayText } from "./status-tray";
 import { changeScreen, renderGameData } from "./game";
-import { ToastNotification } from "./notifications";
+import { ToastNotification } from "./toast-notification";
 import { updateSystemStatusTrayText } from "./system-status-indicator";
 import { createChatMessage } from "./chat";
 import DOMPurify from "dompurify";
